+++ src/status.c
@@ -44,6 +44,8 @@ const char *microlisp_status_string(microlisp_status status) {
         return "MICROLISP_ERR_EVAL_DEPTH";
     case MICROLISP_ERR_PRINT_DEPTH:
         return "MICROLISP_ERR_PRINT_DEPTH";
+    case MICROLISP_ERR_EQUAL_DEPTH:
+        return "MICROLISP_ERR_EQUAL_DEPTH";
     default:
         return "MICROLISP_ERR_UNKNOWN";
     }